requirements.txt
numpy
matplotlib
scikit-learn
torchinfo
puts==0.0.8

